fix(fe): use the current sign-in picker on /mcp and /cli
Both routes mounted AuthWizard in a way that fell back to its legacy
presentation for users with no stored identity.

The wizard detects its framing through `isInsideAuthPanel()`, which was
exported from a markup-less marker component in `ui/` while /mcp and
/cli imported the visual card in `layout/`. Two components with the same
name, only one carrying the context: `inAuthPanel` stayed false, which
dropped the picker's footer styling and hid the switch-mode CTA for a
first-time user. Collapse both into the `layout/` component so the
marker and the card can't drift apart again.

Neither route passed `mode`, so the wizard ran in its `both` default:
generic "Continue with ..." copy, no sign-up CTA, and passkey sign-in
routed through the "Simplify your sign-in" interstitial. Pass
`mode="signin"`, matching /authorize.

The /cli spec asserted the old label was hidden to prove the wizard
wasn't shown; that label no longer exists on the route, so the
assertion would have passed vacuously.
