+++ src/frontend/src/lib/components/wizards/auth/AuthWizard.svelte
@@ -8,7 +8,7 @@
   import SolveCaptcha from "$lib/components/wizards/auth/views/SolveCaptcha.svelte";
   import PickAuthenticationMethod from "$lib/components/wizards/auth/views/PickAuthenticationMethod.svelte";
   import Dialog, { isInsideDialog } from "$lib/components/ui/Dialog.svelte";
-  import { isInsideAuthPanel } from "$lib/components/ui/AuthPanel.svelte";
+  import { isInsideAuthPanel } from "$lib/components/layout/AuthPanel.svelte";
   import SetupOrUseExistingPasskey from "$lib/components/wizards/auth/views/SetupOrUseExistingPasskey.svelte";
   import CreatePasskey from "$lib/components/wizards/auth/views/CreatePasskey.svelte";
   import SystemOverlayBackdrop from "$lib/components/utils/SystemOverlayBackdrop.svelte";
