+++ src/frontend/src/routes/(new-styling)/authorize/+page.svelte
@@ -21,7 +21,7 @@
   import { Trans } from "$lib/components/locale";
   import MigrationIllustration from "$lib/components/illustrations/MigrationIllustration.svelte";
   import Dialog from "$lib/components/ui/Dialog.svelte";
-  import AuthPanel from "$lib/components/ui/AuthPanel.svelte";
+  import AuthPanel from "$lib/components/layout/AuthPanel.svelte";
 
   import RedirectAnimationView from "./views/RedirectAnimationView.svelte";
   import UpgradeSuccessView from "./views/UpgradeSuccessView.svelte";
@@ -555,20 +555,16 @@
       {#if $GUIDED_UPGRADE || $MIN_GUIDED_UPGRADE}
         {@render upgradePanel()}
       {/if}
-      <div
+      <AuthPanel
         class={[
-          "z-1 grid w-full flex-col overflow-hidden rounded-xl px-4 pt-5 pb-8",
-          "max-sm:flex-1",
-          "sm:bg-bg-secondary sm:border-border-secondary sm:border sm:px-6",
+          "z-1",
           ($GUIDED_UPGRADE || $MIN_GUIDED_UPGRADE) &&
             isUpgradeCollapsed &&
             "rounded-t-none",
         ]}
       >
-        <AuthPanel>
-          {@render content()}
-        </AuthPanel>
-      </div>
+        {@render content()}
+      </AuthPanel>
     </div>
   </div>
 {/snippet}
