+++ src/frontend/src/routes/(new-styling)/cli/+page.svelte
@@ -276,7 +276,7 @@
 {:else if phase.kind === "wizard"}
   <div class="flex w-full justify-center max-sm:flex-1 sm:max-w-110">
     <AuthPanel>
-      <AuthWizard {...wizardSignInHandlers}>
+      <AuthWizard {...wizardSignInHandlers} mode="signin">
         <h1 class="text-text-primary my-2 self-start text-2xl font-medium">
           {$t`Sign in`}
         </h1>
